upaded
The .map() method allows you to run a function on each item in the array, returning a new array as the result.

The React useState Hook allows us to track state in a function component.

State generally refers to data or properties that need to be tracking in an application.